split configuration details into two components
split configuration details into two components

add state for editing, remove connector validation
