+++ app/ui-react/packages/ui/src/Customization/apiClientConnectors/detail/ApiConnectorDetailConfigEdit.tsx
@@ -1,0 +1,21 @@
+import { Form } from '@patternfly/react-core';
+import * as React from 'react';
+
+export interface IApiConnectorDetailConfigEdit {
+  // Initial properties
+  properties?: any;
+}
+
+export const ApiConnectorDetailConfigEdit: React.FunctionComponent<IApiConnectorDetailConfigEdit> = ({
+  properties,
+}) => {
+  // tslint:disable:no-console
+
+  return (
+    <>
+      <Form isHorizontal={true} data-testid={'api-connector-details-form'}>
+        <p>Children go here..</p>
+      </Form>
+    </>
+  );
+};
