+++ app/ui-react/packages/ui/src/Customization/apiClientConnectors/detail/index.ts
@@ -1,2 +1,4 @@
-export * from './ApiConnectorDetailCard';
+export * from './ApiConnectorDetailBody';
+export * from './ApiConnectorDetailConfig';
+export * from './ApiConnectorDetailConfigEdit';
 export * from './ApiConnectorDetailHeader';
